Re-factoring tests
